# src/tests/EshopGuard.Api.Tests/Snapshots/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "MeDto": {
        "properties": {
          "displayName": {
            "type": [
              "null",
              "string"
            ]
          },
          "effectiveLocale": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "googleLinked": {
            "type": "boolean"
          },
          "hasPassword": {
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "locale": {
            "type": [
              "null",
              "string"
            ]
          },
          "localeSource": {
            "type": "string"
          },
          "memberships": {
            "items": {
              "$ref": "#/components/schemas/MeMembershipDto"
            },
            "type": "array"
          },
          "pendingInvitations": {
            "items": {
              "$ref": "#/components/schemas/PendingInvitationDto"
            },
            "type": "array"
          }
        },
        "required": [
          "id",
          "email",
          "displayName",
          "locale",
          "effectiveLocale",
          "localeSource",
          "hasPassword",
          "googleLinked",
          "memberships",
          "pendingInvitations"
        ],
        "type": "object"
      },
      "MeMembershipDto": {
        "properties": {
          "legalName": {
            "type": [
              "null",
              "string"
            ]
          },
          "role": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "tenantId": {
            "format": "uuid",
            "type": "string"
          },
          "tenantName": {
            "type": "string"
          }
        },
        "required": [
          "tenantId",
          "tenantName",
          "legalName",
          "role",
          "status"
        ],
        "type": "object"
      },
...
      "MembershipDto": {
        "properties": {
          "role": {
            "type": "string"
          },
          "tenantId": {
            "format": "uuid",
            "type": "string"
          },
          "tenantName": {
            "type": "string"
          }
        },
        "required": [
          "tenantId",
          "tenantName",
          "role"
        ],
        "type": "object"
      },
...
      "PendingInvitationDto": {
        "properties": {
          "expiresAt": {
            "format": "date-time",
            "type": "string"
          },
          "invitationId": {
            "format": "uuid",
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "tenantName": {
            "type": "string"
          }
        },
        "required": [
          "invitationId",
          "tenantName",
          "role",
          "expiresAt"
        ],
        "type": "object"
      },
...
      "SetLocaleRequest": {
        "properties": {
          "locale": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "required": [
          "locale"
        ],
        "type": "object"
      },
      "SetPasswordRequest": {
        "properties": {
          "currentPassword": {
            "type": [
              "null",
              "string"
            ]
          },
          "newPassword": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ],
        "type": "object"
      },
...
      "UpdateMeRequest": {
        "properties": {
          "displayName": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "required": [
          "displayName"
        ],
        "type": "object"
      }
    }
  },
  "info": {
    "title": "EshopGuard.Api | v1",
    "version": "1.0.0"
  },
  "openapi": "3.1.1",
  "paths": {
...
    "/api/me": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "me"
        ],
        "x-problem-codes": [
          "auth.unauthenticated",
          "csrf.invalid",
          "internal_error",
          "rate_limited",
          "request.https_required",
          "request.invalid"
        ]
      },
      "patch": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/UpdateMeRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "me"
        ],
        "x-problem-codes": [
          "auth.unauthenticated",
          "csrf.invalid",
          "internal_error",
          "rate_limited",
          "request.https_required",
          "request.invalid",
          "validation.failed"
        ]
      }
    },
    "/api/me/invitations/{invitationId}/accept": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "invitationId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MembershipDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "me"
        ],
        "x-problem-codes": [
          "auth.unauthenticated",
          "csrf.invalid",
          "internal_error",
          "invitation.email_mismatch",
          "invitation.expired",
          "invitation.not_found",
          "invitation.used",
          "rate_limited",
          "request.https_required",
          "request.invalid"
        ]
      }
    },
    "/api/me/locale": {
      "put": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/SetLocaleRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "me"
        ],
        "x-problem-codes": [
          "auth.unauthenticated",
          "csrf.invalid",
          "internal_error",
          "locale.not_enabled",
          "rate_limited",
          "request.https_required",
          "request.invalid",
          "validation.failed"
        ]
      }
    },
...
    "/api/me/password": {
      "delete": {
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "me"
        ],
        "x-problem-codes": [
          "auth.reauthentication_required",
          "auth.unauthenticated",
          "csrf.invalid",
          "internal_error",
          "rate_limited",
          "request.https_required",
          "request.invalid"
        ]
      },
      "put": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/SetPasswordRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "me"
        ],
        "x-problem-codes": [
          "auth.reauthentication_required",
          "auth.unauthenticated",
          "csrf.invalid",
          "internal_error",
          "password.current_invalid",
          "rate_limited",
          "request.https_required",
          "request.invalid",
          "validation.failed"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "https://localhost/"
    }
  ],
...
}